// pwm-angular/src/app/models/live-news.model.ts
export interface LiveNewsModel {
  ID: string;
  title: string;
  authorID: string;
  description: string;
  content: LiveNewsContentItem[];
  image: string;
  categories: string[];
  createdAt: Date;
  updatedAt: Date;
}

export interface LiveNewsContentItem {
  title: string;
  description: string;
  timestamp: Date | string | any; // Support for Date, string, or Firestore Timestamp
}